アクセス制限の付加

+++ app/controllers/items_controller.rb
@@ -1,4 +1,5 @@
 class ItemsController < ApplicationController
+  before_action :move_to_user_registration, except: [:index, :show]
   before_action :set_item, only: [:edit, :show, :update, :destroy]
 
   # トップページ
@@ -29,8 +30,9 @@ class ItemsController < ApplicationController
     @category = Category.find(params[:id])
   end
 
-  # 商品情報編集ページ
+  # 商品情報編集ページ（編集できるのは出品者であること、かつ、取引が成立していないこと）
   def edit
+    redirect_to item_path(@item) unless @item.user_id == current_user.id && @item.status == 0
   end
   
   def update
@@ -41,8 +43,9 @@ class ItemsController < ApplicationController
     end
   end
 
-  # 商品削除
+  # 商品削除（削除できるのは出品者であること、かつ、取引が成立していないこと）
   def destroy
+    redirect_to root_path unless @item.user_id == current_user.id && @item.status == 0
     if @item.destroy
       redirect_to root_path
     else
@@ -79,5 +82,9 @@ class ItemsController < ApplicationController
 
   def set_item
     @item = Item.find(params[:id])
+  end
+
+  def move_to_user_registration
+    redirect_to new_user_registration_path unless user_signed_in?
   end
 end
